Buys register complete

+++ inventioryApp/src/main/java/com/imperial_net/inventioryApp/dto/ProductUpdateStockDTO.java
@@ -2,7 +2,9 @@ package com.imperial_net.inventioryApp.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductUpdateStockDTO {
-        private Integer stock;
+        private BigDecimal stock;
 }
